refactor(NumberSelectStrategy): 숫자 선택 전락을 별도의 디렉토리로 리펙토링
