+++ src/main/java/lotto/domain/LottoDispenser.java
@@ -2,6 +2,7 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.domain.strategy.NumberSelectStrategy;
 
 public class LottoDispenser {
