+++ src/main/java/lotto/domain/strategy/NumberSelectStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.strategy;
 
 import java.util.List;
