+++ src/main/java/lotto/domain/strategy/RandomNumberSelectStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.strategy;
 
 import java.util.Collections;
 import java.util.List;
