+++ src/test/java/lotto/LottoDispenserTest.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.stream.Stream;
 import lotto.domain.LottoDispenser;
 import lotto.domain.LottoTicketList;
-import lotto.domain.NumberSelectStrategy;
-import lotto.domain.RandomNumberSelectStrategy;
+import lotto.domain.strategy.NumberSelectStrategy;
+import lotto.domain.strategy.RandomNumberSelectStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
